Refactored multi-phasor data windows
