+++ config.py
@@ -1,0 +1,4 @@
+from pathlib import Path
+
+PROJECT_ROOT = Path(__file__).resolve().parent
+UI_DIR = PROJECT_ROOT / "ui files"
